Store media type for files
Since the filename in the filesystem storage will just be the SQL ID, we need to (or would really like to, for performance reasons) somehow store it. This is how.

I considered using memcached or redis or something like that, but decided that, since it's a read-heavy operation and the rest of the SQL row will be needed every time we would need the media type. Also it would be a new service dependency.