Reimplement `OrNull<T>` to use `T: FromStr`
The previous approach used `Deserialize`, but this was problematic because all fields in HTML forms, at least when running it through Axum's `Form` extractor, are strings.
Thus, `FromStr` seems like the best approach, because it is meant for parsing data from strings.

Additionally, document the purpose and the reason for this choice inline with the code.